moved tests to integration
